♻️ 抽离 TopTabBar 组件，MainPanel 保留内容区
